fix a bug in executable creation
The code used eq? to check to see if a given language was in
an list of allowed-to-create-executables languages. But the
language object was passing thru TR and so eq? didn't hold and
thus the check was buggy.
(cherry picked from commit 2d2b3bee2fed3468a317ee2222f58384822f4b8b)