redex: fix `metafunction->pict' width correction for linebreaks
The revised implementation is more clearly and directly expressed
in terms of table cells that span columns.

Also, avoid `panorama' by removing a useless `inset' that seems
likely to be the reason for the `panorama'.

Unix bitmaps to go with the new tests are missing; Robby plans to
add them.