declare failing test should not have failed
This test started failing due to the additions of the calls
to clip, introduced in commit e6dc9c28b492d95.  The actual
difference is only 9 pixels out of about 100,000 relevant
pixels. Even better, here's the precise rgb values for the
mis-matched pixels (the alpha values don't change):

   (224 251 255) vs (229 255 255)
   (200 242 255) vs (205 247 255)
   (205 242 255) vs (210 247 255)
   (205 242 255) vs (210 247 255)
   (205 242 255) vs (210 247 255)
   (205 242 255) vs (210 247 255)
   (200 242 255) vs (205 247 255)
   (224 251 255) vs (229 255 255)
   (242 255 255) vs (247 255 255)

It isn't clear to me why clipping like this would have this
particular change, but I think it is safe to assume that
Redex shouldn't change in response to this!