adjust Redex to disallow certain patterns that are guaranteed never to fail
When you have a named variable that is inside two different
mismatch-named ellipses, e.g.:

  (any_1 ..._!_1 any_1 .._!_1)

then this pattern cannot match anything, since the any_1 constraint
insists that they are bound to the same thing, but the ..._!_1
insists that whatever any_1 is bound to has different lengths
in the two places.

So, make this kind of situation be a syntax error instead.