adjust DrRacket to not compile files in installation's "share"
This rule supplements the old (but now less applicable) rule to
never add debugging-compiled files in the main installation's
"collects" directory.

Note that the new rule does not prevent compilation in
main-distribution packages that are linked, as in the development mode
set up by `make' in the Racet repository's top-level directory.
I think that's probably ok, since the intent of the check is mainly
just to not write to shared installation spaces, which shouldn't be
an issue for development mode.