db/sqlite3: fix memory-management problem
The sqlite3_prepare[_v2]() function takes a byte-string pointer
for a query and returns a pointer into the byte string. That
pointer can be odd-valued, in which case the GC doesn't treat
the pointer as a reference to the byte string (even when the
string is allocated with 'atomic-interior), so arrange for the
original pointer to be retained.