Minimize xrepl-lib dependencies
XREPL already doesn't declare dependencies on its optional
features such as launching DrRacket or opening the docs. The
consensus seems to be that it's best not to declare these and
to gracefully degrade if they're missing. For XREPL, the error
message could be better, but it currently lets the user keep
going when the `dynamic-require` fails.